improve

+++ fullupdate.sh
@@ -1,26 +1,36 @@
 #!/bin/bash
-# download latest rmc dataset
-echo -n "sff "
-source update.sh 
-source ./SFFLoader/credentials.py
+# download latest  dataset
 
-update $sff_download $sff
+loader="SFFLoader"
+
+source update.sh 
+source ./${loader}/credentials.py
+echo ${message} 
+update ${download} ${dataset}
+
+if [ "${success}" =  "no" ]; then
+	echo "end load"
+	exit 1;
+fi
 
 cd ${release}
 eval "$(conda shell.bash hook)"
 conda activate standard
-time python ../SFFLoader/readsdfiles.py
+time python ../${loader}/readsdfiles.py
 
 cd ..
 
 del() {
   shift
   for d in $*; do
-	echo 'removing dataset' $d
-	rm -r "${d}"
+	if [ ! "${d}" == "${release}" ]; then
+	  echo 'removing dataset' $d
+ 	  rm -r "${d}"
+	fi
   done
 
 }
 
 dirs=`ls -dc 2[0-9]*`
 del ${dirs}
+
